Add 'flyctl mysql' as a top-level alias of 'flyctl ext planetscale' and expose the alias in help output
