+++ internal/command/mysql/mysql.go
@@ -1,0 +1,10 @@
+package mysql
+
+import (
+	"github.com/spf13/cobra"
+	"github.com/superfly/flyctl/internal/command/extensions/planetscale"
+)
+
+func New() (cmd *cobra.Command) {
+	return planetscale.New()
+}
